refactor(io): isolate dataset payload types from metadata types
