+++ graphistry/io/dataset_payload_types.py
@@ -1,0 +1,52 @@
+"""
+Server dataset endpoint payload contract types.
+
+Separated from `graphistry.io.types` so Plottable metadata serde types stay
+focused and server response contract evolution is isolated.
+"""
+from typing import Any, Dict, TypedDict
+
+from graphistry.io.types import BindingsDict, EncodingsDict
+from graphistry.validate import URLParamsDict
+
+
+ServerBindingsDict = BindingsDict
+
+
+SIMPLE_ENCODING_SERVER_TO_CLIENT_MAP: Dict[str, str] = {
+    'node_color': 'point_color',
+    'node_size': 'point_size',
+    'node_title': 'point_title',
+    'node_label': 'point_label',
+    'node_icon': 'point_icon',
+    'node_opacity': 'point_opacity',
+    'node_x': 'point_x',
+    'node_y': 'point_y',
+    'edge_color': 'edge_color',
+    'edge_size': 'edge_size',
+    'edge_title': 'edge_title',
+    'edge_label': 'edge_label',
+    'edge_icon': 'edge_icon',
+    'edge_opacity': 'edge_opacity',
+    'edge_source_color': 'edge_source_color',
+    'edge_destination_color': 'edge_destination_color',
+    'edge_weight': 'edge_weight',
+}
+
+
+class DatasetNodeEdgeEncodingPayload(TypedDict, total=False):
+    bindings: ServerBindingsDict
+    complex: Dict[str, Any]
+
+
+class DatasetLegacyPayload(TypedDict, total=False):
+    dataset_id: str
+    bindings: BindingsDict
+    encodings: EncodingsDict
+    metadata: Dict[str, Any]
+    node_encodings: DatasetNodeEdgeEncodingPayload
+    edge_encodings: DatasetNodeEdgeEncodingPayload
+    name: str
+    description: str
+    style: Dict[str, Any]
+    url_params: URLParamsDict
